Revert "Implement a tab bar for the different screens"
