+++ WhiskyBarrl/app/src/main/AndroidManifest.xml
@@ -23,16 +23,23 @@
             android:name=".NewWhiskeyActivity"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar" />
         <activity
+            android:name=".AccountActivity"
+            android:theme="@style/Theme.AppCompat.Light.NoActionBar" />
+        <activity
             android:name=".EditProfile"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar" />
+        <activity android:name=".SearchActivity" />
         <activity
             android:name=".ViewWhiskey"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar" />
         <activity
+            android:name=".MainWhiskeyListActivity"
+            android:theme="@style/Theme.AppCompat.Light.DarkActionBar" />
+        <activity
             android:name=".WhiskeyCommentListActivity"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar" />
         <activity
-            android:name=".MainTabbedActivity"
+            android:name=".UserFavoritesActivity"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar" />
     </application>
