+++ WhiskyBarrl/app/src/main/java/edu/wpi/mis270xteam1/whiskybarrl/LoginActivity.java
@@ -40,7 +40,7 @@ public class LoginActivity extends AppCompatActivity {
             public void onClick(View v) {
                 if (isSuccessfulLogin()) {
                     Toast.makeText(LoginActivity.this, "Login Successful", Toast.LENGTH_SHORT).show();
-                    Intent i = new Intent(LoginActivity.this, MainTabbedActivity.class);
+                    Intent i = new Intent(LoginActivity.this, MainWhiskeyListActivity.class);
                     i.putExtra("username", editTextUsername.getText().toString());
                     startActivity(i);
                 } else {
