+++ WhiskyBarrl/app/src/main/java/edu/wpi/mis270xteam1/whiskybarrl/SearchActivity.java
@@ -1,0 +1,13 @@
+package edu.wpi.mis270xteam1.whiskybarrl;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+
+public class SearchActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_search);
+    }
+}
