--- WhiskyBarrl/app/src/main/java/edu/wpi/mis270xteam1/whiskybarrl/SearchFragment.java
@@ -1,21 +1,0 @@
-package edu.wpi.mis270xteam1.whiskybarrl;
-
-import android.support.v4.app.Fragment;
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-public class SearchFragment extends Fragment {
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_search, container, false);
-        return view;
-    }
-}
