+++ WhiskyBarrl/app/src/main/java/edu/wpi/mis270xteam1/whiskybarrl/ViewWhiskey.java
@@ -1,6 +1,5 @@
 package edu.wpi.mis270xteam1.whiskybarrl;
 
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
@@ -8,9 +7,13 @@ import android.text.TextUtils;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.ListView;
 import android.widget.RatingBar;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class ViewWhiskey extends AppCompatActivity {
     private TextView textViewName;
@@ -21,11 +24,10 @@ public class ViewWhiskey extends AppCompatActivity {
     private EditText commentEditText;
     private Button submitCommentButton;
     private Button viewCommentsButton;
-    private Button favoriteButton;
+    private Button addFavoriteButton;
 
     private int whiskeyId;
     private String currentUsername;
-    private User currentUser;
     private Whiskey whiskey;
     private DatabaseHandler db;
 
@@ -37,7 +39,6 @@ public class ViewWhiskey extends AppCompatActivity {
         currentUsername = getIntent().getStringExtra("username");
         whiskeyId = getIntent().getIntExtra("whiskeyId", -1);
         db = new DatabaseHandler(this);
-        currentUser = db.getUser(currentUsername);
         whiskey = db.getWhiskey(whiskeyId);
 
         textViewName = (TextView) findViewById(R.id.viewWhiskeyNameText);
@@ -48,7 +49,7 @@ public class ViewWhiskey extends AppCompatActivity {
         commentEditText = (EditText) findViewById(R.id.editTextEnterComment);
         submitCommentButton = (Button) findViewById(R.id.commentSubmitButton);
         viewCommentsButton = (Button) findViewById(R.id.viewCommentsListButton);
-        favoriteButton = (Button) findViewById(R.id.addFavoriteButton);
+        addFavoriteButton = (Button) findViewById(R.id.addFavoriteButton);
 
         ratingBar.setFocusable(false);
 
@@ -82,33 +83,7 @@ public class ViewWhiskey extends AppCompatActivity {
             }
         });
 
-        if (db.isFavorite(currentUser, whiskey)) {
-            setFavoriteButtonToDelete();
-        } else {
-            setFavoriteButtonToAdd();
-        }
-    }
-
-    private void populateInformationFields() {
-        textViewName.setText(whiskey.getName());
-        textViewLocation.setText(whiskey.getLocation());
-        textViewAlcCont.setText(Integer.toString(whiskey.getProofLevel()));
-        textViewDescription.setText(whiskey.getDescription());
-        ratingBar.setRating(whiskey.getRating());
-    }
-
-    private void setFavoriteButtonToAdd() {
-        favoriteButton.setText(R.string.add_to_favorites);
-        favoriteButton.setOnClickListener(addFavoriteListener());
-    }
-
-    private void setFavoriteButtonToDelete() {
-        favoriteButton.setText(R.string.remove_from_favorites);
-        favoriteButton.setOnClickListener(deleteFavoriteListener());
-    }
-
-    private View.OnClickListener addFavoriteListener() {
-        return new View.OnClickListener() {
+        addFavoriteButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
                 User user = db.getUser(currentUsername);
@@ -118,27 +93,15 @@ public class ViewWhiskey extends AppCompatActivity {
                         "Added " + whiskey.getName() + " to Favorites",
                         Toast.LENGTH_SHORT
                 ).show();
-
-                // Change the button functionality to remove a favorite.
-                setFavoriteButtonToDelete();
             }
-        };
+        });
     }
 
-    private View.OnClickListener deleteFavoriteListener() {
-        return new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                db.deleteFavorite(currentUser, whiskey);
-                Toast.makeText(
-                        ViewWhiskey.this,
-                        "Removed " + whiskey.getName() + " from Favorites",
-                        Toast.LENGTH_SHORT
-                ).show();
-
-                // Change the button functionality to add a favorite.
-                setFavoriteButtonToAdd();
-            }
-        };
+    private void populateInformationFields() {
+        textViewName.setText(whiskey.getName());
+        textViewLocation.setText(whiskey.getLocation());
+        textViewAlcCont.setText(Integer.toString(whiskey.getProofLevel()));
+        textViewDescription.setText(whiskey.getDescription());
+        ratingBar.setRating(whiskey.getRating());
     }
 }
