+++ WhiskyBarrl/app/src/main/res/layout/activity_account.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".AccountFragment">
+    tools:context=".AccountActivity">
 
 
 
@@ -91,6 +91,12 @@
             android:layout_height="wrap_content"
             android:text="Edit Profile"
             android:id="@+id/buttonEP" />
+
+        <Button
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Your Favorites"
+            android:id="@+id/favoritesButton" />
 
 
     </LinearLayout>
