+++ WhiskyBarrl/app/src/main/res/layout/activity_main_whiskey_list.xml
@@ -7,8 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    android:id="@+id/whiskeyViewLayout"
-    tools:context=".MainWhiskeyListFragment">
+    tools:context=".MainWhiskeyListActivity">
 
     <ListView
         android:layout_width="wrap_content"
@@ -16,4 +15,5 @@
         android:id="@+id/whiskeyListView"
         android:layout_centerVertical="true"
         android:layout_alignParentTop="true" />
+
 </RelativeLayout>
