+++ WhiskyBarrl/app/src/main/res/menu/menu_main_whiskey_list.xml
@@ -2,15 +2,10 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".MainTabbedActivity">
+    tools:context="edu.wpi.mis270xteam1.whiskybarrl.MainWhiskeyListActivity">
 
     <item android:id="@+id/menu_item_new_whiskey"
-        android:icon="@drawable/add_whiskey"
+        android:icon="@android:drawable/ic_menu_add"
         android:title="@string/new_whiskey"
-        app:showAsAction="ifRoom|withText" />
-
-    <item android:id="@+id/menu_item_search_whiskeys"
-        android:icon="@android:drawable/ic_menu_search"
-        android:title="@string/search_whiskeys"
         app:showAsAction="ifRoom|withText" />
 </menu>
