+++ WhiskyBarrl/app/src/main/res/values/strings.xml
@@ -4,10 +4,5 @@
     <string name="action_settings">Settings</string>
     <string name="new_whiskey">Add Whiskey</string>
     <string name="no_comments">No Comments</string>
-    <string name="my_account">My Account</string>
-    <string name="search_whiskeys">Search Whiskeys</string>
-    <string name="add_to_favorites">Add to Favorites</string>
-    <string name="remove_from_favorites">Remove from Favorites</string>
-    <string name="no_whiskeys">No whiskeys are listed.</string>
     <string name="no_favorites">You have not favorited any whiskeys.</string>
 </resources>
